+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* **[ADDED]** `kick` server controller command
 * **[FIXED]** `Zlikavac32\BeanstalkdLibBundle\Command\Runnable\BeanstalkdServerControllerRunnable` crash on empty input line
 
 ## 0.1.0 (2019-05-05)